- fix another float/int conversion
        * platform/graphics/svg/cg/SVGResourceMaskerCg.mm:
        (WebCore::SVGResourceMasker::applyMask): Explicitly cast to int.



Canonical link: https://commits.webkit.org/20545@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@26557 268f45cc-cd09-0410-ab3c-d52691b4dbfc

+++ WebCore/platform/graphics/svg/cg/SVGResourceMaskerCg.mm
@@ -98,7 +98,7 @@ void SVGResourceMasker::applyMask(GraphicsContext* context, const FloatRect& bou
     if (!m_mask)
         return;
     
-    IntSize maskSize(m_maskRect.width(), m_maskRect.height());
+    IntSize maskSize(static_cast<int>(m_maskRect.width()), static_cast<int>(m_maskRect.height()));
 
     // Create new graphics context in gray scale mode for image rendering
     auto_ptr<ImageBuffer> grayScaleImage(ImageBuffer::create(maskSize, true));
